requirements.txt
paramiko
PyYAML
easydict